# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Agent Platform API
  description: |
    Agent Platform Product Experience API - A multi-agent platform for task orchestration 
    and execution with LLM providers, background task processing, workflows, triggers, and connectors.
    
    All endpoints use a standard API response envelope:
    - Success: `{ ok: true, data: ..., requestId: "..." }`
    - Error: `{ ok: false, error: { code: "...", message: "..." }, requestId: "..." }`
  version: "0.7.0-rc.1"
  contact:
    name: Agent Platform Team
servers:
  - url: http://localhost:3003
    description: Development
  - url: http://localhost:3103
    description: E2E Testing
...
paths:
...
  /api/v1/providers:
    get:
      tags: [Agents & Providers]
      summary: List providers
      description: Returns all configured providers for the current user
      operationId: listProviders
      security:
        - sessionAuth: []
      responses:
        "200":
          description: List of providers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Agents & Providers]
      summary: Create provider
      description: Creates a new provider configuration
      operationId: createProvider
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProviderRequest"
      responses:
        "201":
          description: Provider created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/providers/{providerId}:
    patch:
      tags: [Agents & Providers]
      summary: Update provider
      description: Updates a provider configuration
      operationId: updateProvider
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/providerId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProviderRequest"
      responses:
        "200":
          description: Provider updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Agents & Providers]
      summary: Delete provider
      description: Deletes a provider configuration
      operationId: deleteProvider
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/providerId"
      responses:
        "204":
          description: Provider deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/providers/{providerId}/test:
    post:
      tags: [Agents & Providers]
      summary: Test provider connection
      description: Tests the provider connection
      operationId: testProvider
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/providerId"
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestProviderResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/models:
    get:
      tags: [Agents & Providers]
      summary: List models and providers
      description: Returns available providers and optionally session's selected model
      operationId: listModels
      security:
        - sessionAuth: []
      parameters:
        - name: sessionId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Models and providers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: agent-platform-session
      description: Session cookie set after login
    bearerAuth:
      type: http
      scheme: bearer
      description: API token for programmatic access (set API_AUTH_TOKEN env var to enable)
  parameters:
...
    providerId:
      name: providerId
      in: path
      required: true
      schema:
        type: string
      description: Provider ID
...
  schemas:
...
    ApiEnvelope:
      type: object
      required: [ok, requestId]
      properties:
        ok:
          type: boolean
          description: true for success, false for error
        requestId:
          type: string
          description: Traceable request identifier
    ApiErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable error message
        details:
          description: Additional error details
...
    ErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          required: [error]
          properties:
            error:
              $ref: "#/components/schemas/ApiErrorDetail"

    # ---- Auth & Setup ----
...
    ProviderSummary:
      type: object
      required: [providerId, providerType, displayName, enabled, configured, source]
      properties:
        providerId:
          type: string
        providerType:
          type: string
          enum: [openai, openrouter, ollama, custom]
        displayName:
          type: string
        enabled:
          type: boolean
        configured:
          type: boolean
        apiKeyLast4:
          type: string
          nullable: true
        baseUrl:
          type: string
          nullable: true
        selectedModel:
          type: string
          nullable: true
        source:
          type: string
        lastTestStatus:
          type: string
          nullable: true
        lastTestedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProviderListResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/ProviderSummary"
    CreateProviderRequest:
      type: object
      required: [providerType]
      properties:
        providerType:
          type: string
          enum: [openai, openrouter, ollama, custom]
        displayName:
          type: string
        apiKey:
          type: string
        baseUrl:
          type: string
        selectedModel:
          type: string
    UpdateProviderRequest:
      type: object
      properties:
        displayName:
          type: string
        apiKey:
          type: string
        baseUrl:
          type: string
        selectedModel:
          type: string
        enabled:
          type: boolean
    TestProviderResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              type: object
              required: [success, latencyMs]
              properties:
                success:
                  type: boolean
                latencyMs:
                  type: integer
                modelCount:
                  type: integer
                error:
                  type: string
    ModelsResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              type: object
              required: [providers]
              properties:
                providers:
                  type: array
                  items:
                    $ref: "#/components/schemas/ProviderSummary"
                selectedModel:
                  type: string
                selectedProviderId:
                  type: string
...
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            error:
              code: "BAD_REQUEST"
              message: "Invalid request body"
            requestId: "abc12345"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            error:
              code: "UNAUTHORIZED"
              message: "Authentication required"
            requestId: "abc12345"
    Forbidden:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            error:
              code: "FORBIDDEN"
              message: "Access denied"
            requestId: "abc12345"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            error:
              code: "NOT_FOUND"
              message: "Resource not found"
            requestId: "abc12345"
...